requirements.txt
django==4.1
django-debug-toolbar==2.2
django_extensions==3.1.5
psycopg2-binary==2.9.3
python-dotenv==0.19.1
gunicorn==20.1.0
djangorestframework==3.13.1
django-model-utils==4.2.0
djangorestframework==3.13.1
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support

